requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-dotenv==1.0.1
uuid7==0.1.0
